+++ src/main/java/auth/LoginService.java
@@ -16,7 +16,7 @@ public class LoginService {
         return new TokenResponse(accessToken);
     }
 
-    public UserDetails extractMember(String credential) {
+    public UserDetails extractUserDetails(String credential) {
         Long id = Long.parseLong(jwtTokenProvider.getPrincipal(credential));
         Role role = jwtTokenProvider.getRole(credential);
         return new UserDetails(id, role);
