+++ src/main/java/nextstep/config/AuthConfiguration.java
@@ -1,29 +1,30 @@
-package nextstep;
+package nextstep.config;
 
 import auth.LoginController;
 import auth.LoginService;
-import auth.utils.AdminInterceptor;
 import auth.utils.AuthenticationProvider;
 import auth.utils.JwtTokenProvider;
-import auth.utils.LoginMemberArgumentResolver;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @RequiredArgsConstructor
 @PropertySource("classpath:application.yml")
-public class WebMvcConfiguration implements WebMvcConfigurer {
+public class AuthConfiguration implements WebMvcConfigurer {
 
+    @Value("${security.jwt.token.secret-key}")
+    private String secretKey;
+    @Value("${security.jwt.token.expire-length}")
+    private long validityInMilliseconds;
     private final AuthenticationProvider authenticationProvider;
 
     @Bean
     public JwtTokenProvider jwtTokenProvider() {
-        return new JwtTokenProvider();
+        return new JwtTokenProvider(secretKey, validityInMilliseconds);
     }
 
     @Bean
@@ -34,15 +35,5 @@ public class WebMvcConfiguration implements WebMvcConfigurer {
     @Bean
     public LoginController loginController() {
         return new LoginController(loginService());
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider())).addPathPatterns("/admin/**");
-    }
-
-    @Override
-    public void addArgumentResolvers(List argumentResolvers) {
-        argumentResolvers.add(new LoginMemberArgumentResolver(loginService()));
     }
 }
