+++ src/main/java/nextstep/config/WebMvcConfiguration.java
@@ -1,0 +1,30 @@
+package nextstep.config;
+
+import auth.LoginService;
+import auth.utils.AdminInterceptor;
+import auth.utils.JwtTokenProvider;
+import auth.utils.LoginMemberArgumentResolver;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebMvcConfiguration implements WebMvcConfigurer {
+
+    private final JwtTokenProvider jwtTokenProvider;
+    private final LoginService loginService;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new AdminInterceptor(jwtTokenProvider)).addPathPatterns("/admin/**");
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+        argumentResolvers.add(new LoginMemberArgumentResolver(loginService));
+    }
+}
